+++ crates/block-producer/src/test_utils/block.rs
@@ -1,8 +1,9 @@
-use std::iter;
-
 use miden_objects::{
     batch::ProvenBatch,
-    block::{Block, BlockAccountUpdate, BlockHeader, BlockNoteIndex, BlockNoteTree, NoteBatch},
+    block::{
+        BlockAccountUpdate, BlockHeader, BlockNoteIndex, BlockNoteTree, OutputNoteBatch,
+        ProvenBlock,
+    },
     crypto::merkle::{Mmr, SimpleSmt},
     note::Nullifier,
     transaction::OutputNote,
@@ -10,10 +11,6 @@ use miden_objects::{
 };
 
 use super::MockStoreSuccess;
-use crate::{
-    block::BlockInputs,
-    block_builder::prover::{block_witness::BlockWitness, BlockProver},
-};
 
 /// Constructs the block we expect to be built given the store state, and a set of transaction
 /// batches to be applied
@@ -71,31 +68,6 @@ pub async fn build_expected_block_header(
     )
 }
 
-/// Builds the "actual" block header; i.e. the block header built using the Miden VM, used in the
-/// node
-pub async fn build_actual_block_header(
-    store: &MockStoreSuccess,
-    batches: Vec<ProvenBatch>,
-) -> BlockHeader {
-    let updated_accounts: Vec<_> =
-        batches.iter().flat_map(|batch| batch.account_updates().iter()).collect();
-    let produced_nullifiers: Vec<Nullifier> =
-        batches.iter().flat_map(ProvenBatch::produced_nullifiers).collect();
-
-    let block_inputs_from_store: BlockInputs = store
-        .get_block_inputs(
-            updated_accounts.iter().map(|(&account_id, _)| account_id),
-            produced_nullifiers.iter(),
-            iter::empty(),
-        )
-        .await
-        .unwrap();
-
-    let (block_witness, _) = BlockWitness::new(block_inputs_from_store, &batches).unwrap();
-
-    BlockProver::new().prove(block_witness).unwrap()
-}
-
 #[derive(Debug)]
 pub struct MockBlockBuilder {
     store_accounts: SimpleSmt<ACCOUNT_TREE_DEPTH>,
@@ -103,7 +75,7 @@ pub struct MockBlockBuilder {
     last_block_header: BlockHeader,
 
     updated_accounts: Option<Vec<BlockAccountUpdate>>,
-    created_notes: Option<Vec<NoteBatch>>,
+    created_notes: Option<Vec<OutputNoteBatch>>,
     produced_nullifiers: Option<Vec<Nullifier>>,
 }
 
@@ -140,7 +112,7 @@ impl MockBlockBuilder {
     }
 
     #[must_use]
-    pub fn created_notes(mut self, created_notes: Vec<NoteBatch>) -> Self {
+    pub fn created_notes(mut self, created_notes: Vec<OutputNoteBatch>) -> Self {
         self.created_notes = Some(created_notes);
 
         self
@@ -153,7 +125,7 @@ impl MockBlockBuilder {
         self
     }
 
-    pub fn build(self) -> Block {
+    pub fn build(self) -> ProvenBlock {
         let created_notes = self.created_notes.unwrap_or_default();
 
         let header = BlockHeader::new(
@@ -170,28 +142,27 @@ impl MockBlockBuilder {
             1,
         );
 
-        Block::new(
+        ProvenBlock::new_unchecked(
             header,
             self.updated_accounts.unwrap_or_default(),
             created_notes,
             self.produced_nullifiers.unwrap_or_default(),
         )
-        .unwrap()
     }
 }
 
 pub(crate) fn flatten_output_notes<'a>(
-    batches: impl Iterator<Item = &'a NoteBatch>,
+    batches: impl Iterator<Item = &'a OutputNoteBatch>,
 ) -> impl Iterator<Item = (BlockNoteIndex, &'a OutputNote)> {
     batches.enumerate().flat_map(|(batch_idx, batch)| {
-        batch.iter().enumerate().map(move |(note_idx_in_batch, note)| {
-            (BlockNoteIndex::new(batch_idx, note_idx_in_batch).unwrap(), note)
+        batch.iter().map(move |(note_idx_in_batch, note)| {
+            (BlockNoteIndex::new(batch_idx, *note_idx_in_batch).unwrap(), note)
         })
     })
 }
 
 pub(crate) fn note_created_smt_from_note_batches<'a>(
-    batches: impl Iterator<Item = &'a NoteBatch>,
+    batches: impl Iterator<Item = &'a OutputNoteBatch>,
 ) -> BlockNoteTree {
     let note_leaf_iterator =
         flatten_output_notes(batches).map(|(index, note)| (index, note.id(), *note.metadata()));
@@ -201,6 +172,8 @@ pub(crate) fn note_created_smt_from_note_batches<'a>(
 
 pub(crate) fn block_output_notes<'a>(
     batches: impl Iterator<Item = &'a ProvenBatch> + Clone,
-) -> Vec<Vec<OutputNote>> {
-    batches.map(|batch| batch.output_notes().to_vec()).collect()
+) -> Vec<OutputNoteBatch> {
+    batches
+        .map(|batch| batch.output_notes().iter().cloned().enumerate().collect())
+        .collect()
 }