+++ crates/store/src/db/mod.rs
@@ -11,7 +11,7 @@ use miden_node_proto::{
 };
 use miden_objects::{
     account::{AccountDelta, AccountId},
-    block::{Block, BlockHeader, BlockNoteIndex, BlockNumber},
+    block::{BlockHeader, BlockNoteIndex, BlockNumber, ProvenBlock},
     crypto::{hash::rpo::RpoDigest, merkle::MerklePath, utils::Deserializable},
     note::{NoteId, NoteInclusionProof, NoteMetadata, Nullifier},
     transaction::TransactionId,
@@ -87,7 +87,12 @@ impl NoteRecord {
     /// ensure ordering is correct.
     fn from_row(row: &rusqlite::Row<'_>) -> Result<Self> {
         let block_num = read_block_number(row, 0)?;
-        let note_index = BlockNoteIndex::new(row.get(1)?, row.get(2)?)?;
+        let batch_idx = row.get(1)?;
+        let note_idx_in_batch = row.get(2)?;
+        // SAFETY: We can assume the batch and note indices stored in the DB are valid so this
+        // should never panic.
+        let note_index = BlockNoteIndex::new(batch_idx, note_idx_in_batch)
+            .expect("batch and note index from DB should be valid");
         let note_id = row.get_ref(3)?.as_blob()?;
         let note_id = RpoDigest::read_from_bytes(note_id)?;
         let note_type = row.get::<_, u8>(4)?.try_into()?;
@@ -454,7 +459,7 @@ impl Db {
         &self,
         allow_acquire: oneshot::Sender<()>,
         acquire_done: oneshot::Receiver<()>,
-        block: Block,
+        block: ProvenBlock,
         notes: Vec<(NoteRecord, Option<Nullifier>)>,
     ) -> Result<()> {
         self.pool
@@ -469,7 +474,7 @@ impl Db {
                     &transaction,
                     &block.header(),
                     &notes,
-                    block.nullifiers(),
+                    block.created_nullifiers(),
                     block.updated_accounts(),
                 )?;